an error a day will surely keep me awake

+++ src/RageUtil.cpp
@@ -28,7 +28,7 @@ float RandomFloat( int &seed )
 	if( seed < 0 )
 			seed += IM;
 
-	const float AM = .999999f / IM;
+	const float AM = .999999f / (float) IM;
 	float ans = AM * seed;
 
 	seed ^= MASK;
